FROMLIST: arm: dts: k3-am62p5-sk-u-boot: Add boot phase tag for USB0
The USB0 instance of USB on AM62Px SoC can be used for USB DFU boot. This
requires USB0 to be enabled at all stages of the boot process. In order to
support USB DFU boot on AM62P5-SK, add the "bootph-all" property to USB0.

Since U-Boot device-tree is updated via DT-Sync with the Linux device-tree,
the link to the equivalent Linux device-tree patch is included below.

Link: https://patch.msgid.link/20250122124223.1118789-3-s-vadapalli@ti.com
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>