PENDING: include: env: ti: prioritize scripting over EFI
There are a few issues with the current stdboot flow that are
complicating the SystemReady and EFI transition. For example how to load
overlays on a unknown kernel version. How does the kernel gracefully
handle the older device tree U-Boot gives it. The list goes on.

However the current pre-compiled hush scripts remain complicated,
brittle and a sore point for anyone not using Yocto/Arago for building
their rootfs and wishes to use one of these stdboot functions.

So in the hopes of maximizing the number of people possible (I'm
expecting a little push back here ;)) lets move the 'script' bootmethod
first in line in the scan so any distribution which injects a valid
boot.scr.uimg file into their / or /boot folders will be used to boot.
If these are missing or invalid we can fallback to efi and extlinux

Right now Arago does not inject a boot.scr.uimg file so this patch will
have no change in how K3 devices currently boot.

Signed-off-by: Bryan Brattlof <bb@ti.com>