PENDING: HACK: board: ti: j721e: evm: Explicitly probe SERDES driver for EthFw
Commit 434f84b74983 ("board: ti: j721e: Drop SERDES PHY init from board
file") removed SERDES driver probe for both the Sierra SERDES driver which
is applicable to J721E and the Torrent SERDES driver which is applicable
to J7200.

While the Torrent SERDES driver conforms to the UCLASS_PHY framework and
will therefore be probed if the device-tree nodes indicate the desired
SERDES configuration, the Sierra SERDES driver belongs to UCLASS_MISC
framework, due to which it needs to be explicitly probed for the SERDES
to be configured as described by the device-tree nodes.

Hence, probe the Sierra SERDES driver explicitly and mark this as a HACK
since the correct fix is that of updating the Sierra SERDES driver and
modelling it as UCLASS_PHY driver.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>