TI: HACK: usb: cdns3: continue probe even when USB PHY device does not exist
Prior to commit cd295286c786 ("usb: cdns3: avoid error messages if phys
don't exist"), cdns3_probe() errors out only on failing to initialize the
USB2/USB3 PHY. However, since commit cd295286c786, absence of the PHY's
device driver is also treated as an error, resulting in a regression.

Extend commit cd295286c786 to treat -ENODEV as an acceptable return value
of generic_phy_get_by_name() and continue cdns3_probe() as was the case
prior to the commit.

ENODEV is acceptable in the special case where the previous stage of the
boot process has already configured the USB PHY, due to which it can be
used without the need for re-configuring it. This however is a HACK since
the correct approach is that of probing the driver for the USB PHY and
figuring out in the driver that the PHY has already been configured.

Fixes: cd295286c786 ("usb: cdns3: avoid error messages if phys don't exist")
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: MD Danish Anwar <danishanwar@ti.com>