TI: HACK: include/scmi_protocols: increase name size
Newer protocol versions of SCMI have increased the size of these
structures so increase them here to prevent buffer overruns.

I'm labeling this a HACK as we're not upgrading all messages, just the
ones we need for the 62L to function properly. When porting the 62L to
U-Boot we will need to upgrade the SCMI version.

Signed-off-by: Bryan Brattlof <bb@ti.com>