TI: bootstd: android: add support of init_boot
init_boot is part of the generic boot partition [1], this partition
contains the generic ramdisk.
This partition is optional and the support is available only with boot
image v4.
If init_boot is not present on the device with boot image v4, the
ramdisk should be placed in the boot image.

[1] https://source.android.com/docs/core/architecture/partitions/generic-boot

Signed-off-by: Julien Masson <jmasson@baylibre.com>
Signed-off-by: Mattijs Korpershoek <mkorpershoek@baylibre.com>