PENDING: arch: mach-k3: Kconfig: Add symbols for IO+DDR Low Power Mode
There is one new symbol to enable support to resume from IO+DDR where
the SoC is turned off and DDR is in self-refresh.

The other symbol is the address of a memory region to be used to store
meta data. This has to be the same address as is used by the firmware.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Signed-off-by: Kendall Willis <k-willis@ti.com>
Acked-by: Dhruva Gole <d-gole@ti.com>