PENDING: arm: mach-k3: Add config option for booting HSM core
Some K3 SoCs like J721S2, and J784S4 have a HSM (High Security Module)
M4F core in the Wakeup Voltage Domain which could be used to run secure
services like Authentication. Add a config option which indicates that
the SoC has a HSM core and the firmware for the core is available. The
firmware for HSM core is packed in the tispl.bin fit image at A72 SPL
stage. So, Enable the config in applicable J7* R5 and A72 defconfigs.

Signed-off-by: Beleswar Padhi <b-padhi@ti.com>
Reviewed-by: Udit Kumar <u-kumar1@ti.com>