PENDING: arm: dts: k3-binman: Add template for packing HSM firmware
The HSM M4 core needs to be booted at R5 SPL stage so that it can be
used for further Authentication and security services. Pack the HSM
firmware in tispl.bin fit image so that it can be unloaded and used by
R5 SPL to boot the HSM core. The size of HSM firmware is padded upto 256
KB. The HSM binaries are unloaded at a temporary DDR address.

Signed-off-by: Beleswar Padhi <b-padhi@ti.com>