PENDING: arm: mach-k3: r5: common: Add support to boot HSM M4 core
The tispl.bin fit image is packed with the HSM firmware image. Populate
the "os" info of the image so that it can be detected and loaded at
corresponding DDR address. Further, invoke the load and boot of HSM core
at R5 SPL stage. Boot flow for HSM M4 core is as below:

1. Request control of HSM M4F remote processor.
2. Assert Reset on the HSM M4F remote processor.
3. For HS devices, Request Secure Entity to Authenticate and Load HSM
   firmware into core's internal SRAM memory region. For GP devices,
   load the unsigned firmware manually.
4. Deassert Reset on the HSM M4F remote processor.
5. Release control of HSM M4F remote processor.

Signed-off-by: Beleswar Padhi <b-padhi@ti.com>