PENDING: remove any firewalls ROM has configured for HSRAM
ROM will configure a firewall to only allow HSRAM to be touched by the
R5 core. Any outside entity like DMA will not have access to this region
however this can be problematic when U-Boot, running on the A53 loads
firmware that runs out of this region.

To simplify things remove the firewall here and let the remote core
firmware place a new firewall themselves if they wish.

Signed-off-by: Bryan Brattlof <bb@ti.com>