PENDING: board: ti: Fix incorrect ordering of the otp_config fields
The security board configuration contains a `otp_config` structure for
controlling the access to the OTP rows. As per the TISCI documentation,
this structure defines two fields `otp_entry` and `write_host_id` in
that order for defining the read and write access to the OTP rows.

The order is, however, reversed in the secure board configuration of
certain devices. This leads to the incorrect parsing of the `otp_config`
structure ultimately resulting in the read and write failure.

Fix the order for the devices where it was incorrect.

Signed-off-by: Prashant Shivhare <p-shivhare@ti.com>
Acked-by: Vishal Mahaveer <vishalm@ti.com>