PENDING: bloblist: Move bloblist not found error to debug message
CONFIG_BLOBLIST_FIXED_ADDRESS serves two scenarios. In the first, the
first-stage U-Boot bootloader can retrieve the bloblist from the
previous stage at this address if present. Otherwise, the first-stage
U-Boot bootloader can allocate a bloblist structure at this address and
pass it on to the next stage.

Commit 66131310d8ff ("bloblist: Load the bloblist from the previous
loader") assumes only the first scenario that the first-stage U-Boot
bootloader should retrieve the bloblist from a previous stage
bootloader and generates the following warning if the bloblist is not
found at the CONFIG_BLOBLIST_FIXED_ADDRESS set in the defconfig:

"Bloblist at 80d00000 not found (err=-2)"

This does not account for the second scenario. For example, TI K3
platforms do not use a bloblist before the first-stage U-Boot bootloader
but still use CONFIG_BLOBLIST_FIXED_ADDRESS to instruct the first-stage
U-Boot bootloader (i.e., A53 SPL) to allocate a bloblist structure at
this address and pass it on to the next stage.

To accommodate both scenarios, move the warning message to a debug log.

Fixes: 66131310d8ff ("bloblist: Load the bloblist from the previous loader")
Signed-off-by: Devarsh Thakkar <devarsht@ti.com>