UPSTREAM: board: ti: am335x: Fix selection of BeagleBone Green
commit f364ea8c2571540c10ecc9fab456760936eb7fc6 upstream.

The BeagleBone Green board is a revision of the BoneBlack board.
Having BeagleBone Black devicetree listed before BeagleBone Green will
select always the BeagleBone Black devicetree following the functioning of
board_fit_config_name_match().

Fix it by changing the test condition and selecting BoneBlack board only
if it is not a revision of this board.

Signed-off-by: Kory Maincent <kory.maincent@bootlin.com>