PENDING: ram: ti: k3-ddrss: Add workaround sequence for errata i2487
As per errata
i2487: "the low power modes may inadvertently corrupt DDR contents"

Based on the recommended SW workaround sequence to
clear data retention latch on every boot and/or abort scenarios to
make sure that latch is initialized to zero rather than unknown value.

Executing this sequence during boot helps to successfully resume the SoCs
from IO+DDR low power mode.
This sequence is executed in partial I/O resume as well as all the
other conditions when reset is issued during low power mode entry/exit
or when the SoC is in idle mode.

The sequence should be executed for SoCs that support IO+DDR mode, that
is AM62AX and AM62PX.

Signed-off-by: Akashdeep Kaur <a-kaur@ti.com>
Reviewed-by: Kendall Willis <k-willis@ti.com>
Reviewed-by: Dhruva Gole <d-gole@ti.com>