PENDING: configs: add defconfigs for am6254xxl
OP-TEE for current K3 devices resides from 0x9e800000 to 0xa0000000
which needs to be moved to 0x80080000 to free up space at end of DDR in
systems with 512MiB of memory. This is required to allow U-Boot to
relocate to end of DDR before booting to the kernel.

Therefore defconfigs for AM6254xxl include the respective existing AM62x
configs with updated dt for r5 and modified memory map for a53 SPL to
account for the new OP-TEE address.

Also make Bryan Brattlof and Tom Rini the maintainers for AM6254xxl SIP.

Signed-off-by: Anshul Dalal <anshuld@ti.com>