FROMLIST: Makefile: Fix of_list_dirs with EXT_DTB
EXT_DTB overrides the of_list_dirs that is passed to binman during
builds. This causes binman to not find the normal DTB paths and instead
it just is able to use the path for EXT_DTB only.

Add the default of_list_dirs that is used for normal binman builds to
EXT_DTB usecase.

Link: https://lore.kernel.org/r/20250626-b4-upstream-fix-icssg-fit-v1-1-95eff1c853a4@ti.com
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>