FROMLIST: spl: Kconfig: allow disabling fallback during os boot
During falcon boot in FS mode, if the SPL fails to load the payload
(kernel image and args), the execution fallbacks to standard u-boot boot
mode.

This might not be desired in cases when u-boot payload (tifalcon.bin in
our case) is dysfuncitonal by itself or user's access to standard u-boot
should be restricted (for security reasons for example).

In our falcon use case, this fallback allows the user to bypass
authenticated boot by just deleting the fitImage which causes SPL to
reload tifalcon.bin. The boot proceeds as usual until ATF jumps to
PRELOADED_BL33_BASE where the kernel has not been loaded causing the
core to execute whatever else was written at that address.

Therefore this patch adds a new config symbol SPL_FALCON_ALLOW_FALLBACK
that explicitly toggles the fallback for MMC FS boot. It is set to 'y'
by default so as to not break any existing falcon mode systems.

Signed-off-by: Anshul Dalal <anshuld@ti.com>