PENDING: dma: ti: k3-udma-common: fix dma_addr_t typecasts
With the change to dma_addr_t from u32 to u64 for the R5 SPL, the
existing typecasts from void* to int or vice-versa cause the compiler to
throw a "cast from pointer to integer of different size".

Therefore this patch changes the casts to a uintptr_t which is
guaranteed to hold any pointer value thus supressing the compiler
warning.

Signed-off-by: Anshul Dalal <anshuld@ti.com>