TI: Add CONFIG_AVB_PUBKEY_FILE to specify the root avb public key
The new config CONFIG_AVB_PUBKEY_FILE is used to specify the path to the
public key file that u-boot uses as the root public key. Previously,
the root public key was hard coded in avb_verify.c and therefore was
non-trivial to replace it with a custom key.

The public key that was hard-coded is extracted as default.avbpubkey
and is used as the default.

Bug: 181924266
Test: DIST_DIR=out/dist BUILD_CONFIG=u-boot/build.config.crosvm.aarch64
build/build.sh -j80

Signed-off-by: Jiyong Park <jiyong@google.com>
(cherry picked from commit e609074ced3dd07ca2a13d4becbae997e3a48c9f)
Link: https://android.googlesource.com/platform/external/u-boot/+/e609074ced3dd07ca2a13d4becbae997e3a48c9f
Signed-off-by: Guillaume La Roque (TI.com) <glaroque@baylibre.com>
[mkorpershoek: fixed conflict around _SPL_TPL_ support]
Signed-off-by: Mattijs Korpershoek (TI.com) <mkorpershoek@baylibre.com>
Signed-off-by: Maxime Fournier (TI.com) <mfournier@baylibre.com>