FROMLIST: board: ti: Fix CAPE_EEPROM_BUS_NUM Kconfig dependency
The CAPE_EEPROM_BUS_NUM configuration option was incorrectly depending
on CMD_EXTENSION, which represents the extension board command. However,
the cape scan functionality can be built and used independently of the
command interface through the SUPPORT_EXTENSION_SCAN option.

Change the dependency from CMD_EXTENSION to SUPPORT_EXTENSION_SCAN to
properly reflect that the I2C bus configuration is needed for the cape
scan function itself, not specifically for the command.

Link: https://patchwork.ozlabs.org/project/uboot/patch/20251013-feature_sysboot_extension_board-v3-3-04067b7422a5@bootlin.com/
Signed-off-by: Kory Maincent (TI.com) <kory.maincent@bootlin.com>