FROMLIST: boot: Add UCLASS support for extension boards
Introduce UCLASS-based extension board support to enable more
standardized and automatic loading of extension board device tree
overlays in preparation for integration with bootstd and pxe_utils.

Several #if CONFIG_IS_ENABLED are used in cmd/extension_board.c to ease the
development but don't worry they are removed later in the series.

Link: https://patchwork.ozlabs.org/project/uboot/patch/20251013-feature_sysboot_extension_board-v3-8-04067b7422a5@bootlin.com/
Signed-off-by: Kory Maincent (TI.com) <kory.maincent@bootlin.com>