FROMLIST: distro_bootcmd: Add support for customizable find_distro_rootpart
Commit d0ba0ca45a49 ("distro_bootcmd: Set distro_bootpart_uuid for block
devices") added support for the distro_boot_part_uuid environment variable
to allow using PARTUUID in the kernel command line.

However, the way it was written only supports the case where the boot files
and the rootfs are located in the same partition. There are many cases
where the boot partition and rootfs partition are separate.

Add a customizable find_distro_rootpart command to allow each board to
define its own logic for locating the rootfs partition. This provides
flexibility for boards with custom partition layouts while maintaining
backward compatibility for the default case where boot and rootfs share
the same partition.

Link: https://patchwork.ozlabs.org/project/uboot/patch/20251031-feature_distro_rootpart-v1-1-4e7c465708e8@bootlin.com/
Signed-off-by: Kory Maincent (TI.com) <kory.maincent@bootlin.com>