PENDING: arm: dts: k3-j722s-binman: Add path to HSM firmware blob
The J722S SoC has an HSM M4F core in the WKUP domain. While support
for loading the HSM core was already added, the k3-j722s-binman.dtsi
file did not specify the path to the HSM firmware. As a result, the HSM
firmware was not being included in tispl.bin even when the
CONFIG_K3_HSM_FW option was enabled.

Add the missing path to the HSM firmware binary in the J722S binman
description and updates the list of loadable binaries to include the HSM
firmware when the configuration is enabled.

Fixes: dc6870d4a04 ("PENDING: arm: dts: k3-binman: Add template for packing HSM firmware")
Signed-off-by: Beleswar Padhi <b-padhi@ti.com>