PENDING: arm: dts: k3-binman: Add template for signing HSM firmware blob
With the new system firmware update, the HSM firmware needs to be signed
before invoking TIFS to authenticate and load the blob to HSM core. Add
necessary fields needed in the template which will be populated in the
boot extension and load extension in the x509 certificate for HSM
binary.

Signed-off-by: Beleswar Padhi <b-padhi@ti.com>