PENDING: configs: j784s4_evm_r5: Used main_r5 reserved memory in DDR for malloc heap
When resuming from suspend to ram, we load again the DM firmware.
However, we have to be sure to not modify the memory used by
Linux.

Currently the SPL malloc region is in DDR was in place that could be used
by Linux. Instead of it use a memory address that is located in a
reserved location of main_r5ss core0 dma memory region.

Linux will reload the firmwares and these reserved regions are invalid
once the system goes to suspend.

At resume we need power domain driver to enable R5 cores.
So, enable powerdomain driver in R5.

Signed-off-by: Gregory CLEMENT <gregory.clement@bootlin.com>
Co-developed-by: Prasanth Babu Mantena <p-mantena@ti.com>
Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>