PENDING: ram: k3-ddrss: Add exit retention support
Add the exit retention support

On J7200 and J784s4, the DDR resume sequence is:
- exit DDR from retention
- de-assert the DDR_RET pin
- restore DDR max frequency
- exit DDR from low power

On J7200, it can be done in one function since it has only one DDR.
But for j784s4, as it has 4 DDRs with 4 controllers.
We have to separate each action in order to de-assert the DDR_RET pins
at the right time, because the DDR_RET pins are all tied together
in hardware.
Additional interleaving to be taken care and configuring of PMIC differs
as we have one PMIC on j784s4 and two on J7200.

Signed-off-by: Thomas Richard <thomas.richard@bootlin.com>
Signed-off-by: Gregory CLEMENT <gregory.clement@bootlin.com>
Signed-off-by: Richard Genoud <richard.genoud@bootlin.com>
Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>