PENDING: arm: dts: k3-j784s4: binman: Add new configuration for LPM
This configuration addition is needed when the board is on its resume path.
DM firmware is only parsed and loaded from the flash to its address space.
All the other needed components are present in the DDR while resume.
So, this is only needed to be selected at the resume path.

Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>