PENDING: j784s4: init: Enable DDR_RET exit for j742s2
On the resume path, we perform an exit from the DDR self refresh state.
The self-refresh exit sequence is performed in a loop for each DDR controller.

The J784S4 has 4 ddr controllers and J742S2 has 2 controllers.

Select the number of ddr controllers on the basis of
CONFIG_TARGET_J742S2_R5_EVM build flag.

Signed-off-by: Abhash Kumar Jha <a-kumar2@ti.com>
Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>