PENDING: arm: dts: k3-j7200: Extend firewall for ATF region to TIFS
Extend the access to SRAM region of ATF to TIFS as well. This is
needed for TIFS for encryption and decryption of ATF as a part of
low power mode sequence. TIFS encrypts the ATF while entering into
low power mode and decrypts it back while resuming back.
So, giving permissions for TIFS to access this region.

Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>
Acked-by: Manorit Chawdhry <m-chawdhry@ti.com>