TI: board: ti: am64x: Add on-board clkgen overlay to environment
Revisions B, C and D of the AM642-EVM have an on-board clock generator
that drives the reference clock to the PCIe connector. To avoid contention
between the on-board clock generator driven reference clock and the SERDES
generated reference clock, the device-tree overlay named
"k3-am642-evm-onboard-clkgen-pcie-serdes.dtbo" has to be applied to update
the device-tree passed to Linux.

Hence, add support for automatically adding the device-tree overlay to the
'name_overlays' environment variable for EVM revisions B, C and D.

Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: MD Danish Anwar <danishanwar@ti.com>