UPSTREAM: arm: dts: k3-am625-verdin-r5*: Remove clocks from main_timer0
commit 2696d3baea5910c9fb9f93d14e2354eed05f4e15 upstream.

The commit 79d91e77f4c2 ("clk: ti: clk-k3-pll: Add additional robustness
steps to the PLL sequence") introduced a change which requires the
main_timer0 to not rely on it's own clocks which anyways was wrong.

Fix it by removing the clock dependency for it, also while at it, move
it from u-boot.dtsi to R5 as that is the only entity that should require
it as DM isn't up.

Fixes: 5d1aac358f3c ("arm: dts: k3-*-r5: Remove clocks from mcu_timer0")
Acked-by: Francesco Dolcini <francesco.dolcini@toradex.com>
Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>