PENDING: configs: j784s4_evm_r5: Disable SPL_MULTI_DTB_FIT for j784s4
Remove SPL_MULTI_DTB_FIT from j784s4_r5_defconfig. There are seperate
defconfigs for j784s4 and am69 devices. This config is no longer needed.
This was an issue with the resume flow, when we are selectively picking
the lpm config, which has info of DM image only. The path of resume will
be broken when this config is enabled. Disabling this as we don't need.

Signed-off-by: Prasanth Babu Mantena <p-mantena@ti.com>