TI: HACK: arm: dts: k3-j784s4-evm-u-boot: Use USB 2.0 PHY for DFU Boot
Until commits [0] and [1], the SERDES driver wasn't enabled and USB DFU
boot proceeded with the USB 2.0 PHY. Since [0] and [1], the SERDES driver
attempts to configure the SERDES but fails. Since the device-tree node for
USB0 specifies SERDES as the USB 3.0 PHY, the USB driver attempts to use
it as the USB PHY. SERDES configuration fails and causes USB DFU Boot to
fail as well. Hence, as a HACK, update the USB0 device-tree node to
ensure that USB 2.0 PHY is used for USB DFU Boot. This is until the
root-cause for SERDES configuration failure is identified and fixed.

[0]: commit bab9d338069f ("FROMLIST: configs: j784s4_evm_r5_defconfig: Enable configs for PCIe boot")
[1]: commit fe10c6280eea ("FROMLIST: configs: j784s4_evm_a72_defconfig: Enable configs for PCIe boot")
Signed-off-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Reviewed-by: Udit Kumar <u-kumar1@ti.com>